pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "backstage-server-lab"
version = "0.1.1"
description = "Single-node lab workspace for Vast.ai/Jupyter with minimal MLOps spine"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "jupyterlab>=4.2.0",
  "kaggle>=1.7.4",
  "mlflow>=2.18.0",
  "networkx>=3.3",
  "numpy>=2.1.0",
  "orjson>=3.10.0",
  "platformdirs>=4.3.0",
  "tensorboard>=2.18.0",
  "pandas>=2.2.0",
  "pyyaml>=6.0.1",
  "psutil>=6.0.0",
  "rich>=13.9.0",
  "streamlit>=1.40.0",
  "typer>=0.12.0",
  "torch>=2.5.0"
]

[project.optional-dependencies]
dev = [
  "ruff>=0.7.0",
  "pytest>=8.3.0"
]

[project.scripts]
labops = "labops.cli:main"
kaggle-mashup = "labops.kaggle_mashup_app:main"

[tool.hatch.build.targets.wheel]
packages = ["src/labops"]
